fix: adapt tests involving kt-math to new API from v0.6.2

+++ test-solve/src/commonTest/kotlin/it/unibo/tuprolog/solve/stdlib/function/FloatIntegerPartTest.kt
@@ -44,7 +44,7 @@ internal class FloatIntegerPartTest {
             val (integerPart, _) = parts
             assertEquals(
                 integerPart,
-                FloatIntegerPart.computeOf(Numeric.of(input)).castToReal().value.toDouble()
+                FloatIntegerPart.computeOf(Numeric.of(input)).castToReal().value
             )
         }
     }
